Move services to a separate directory
